Renamed files CTiglBSplineFitMod to CTiglBSplineFit
